+++ backend/src/main/java/wooteco/retrospective/exception/NotFoundException.java
@@ -1,4 +1,4 @@
-package wooteco.exception;
+package wooteco.retrospective.exception;
 
 public class NotFoundException extends RetrospectiveException {
