+++ backend/src/main/java/wooteco/retrospective/exception/RetrospectiveAdvice.java
@@ -1,4 +1,4 @@
-package wooteco.exception;
+package wooteco.retrospective.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.MethodArgumentNotValidException;
