+++ backend/src/test/java/wooteco/retrospective/infrastructure/dao/member/MemberDaoTest.java
@@ -1,4 +1,4 @@
-package wooteco.retrospective.dao.member;
+package wooteco.retrospective.infrastructure.dao.member;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
